backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
openai==1.35.7
PyMySQL==1.1.1
python-dotenv==1.0.1
